lint the main PHP file according to PHP coding standards and other simple cleanups
A couple comments:

The Version line in the main PHP file did not match the published plugin version.

The Domain Path line is needed in the plugin header because the plugin is using a custom directory for translatable strings file.

The way I have the call to the load_plugin_textdomain function is the currently recommended way of using that function.

The "css_plus_plugin_func" function had a nested <p> element within another <p> element, which is invalid HTML.

The "css_plus_engine" function needed to be fixed this way because it would throw a syntax error on pages which did not have the global $post variable initialized, such as 404 pages.