migrate the plugin's functions into the class
This change reduces the number of functions in the global scope and makes the plugin more modular and manageable.

The save_post function now does a couple EXTREMELY IMPORTANT authorization checks: the correct nonce must be set, and the user making the change must have an edit_page capability on the post. This is where such authorization must be verified, not in the function which outputs the meta box, since there the WP core itself checks whether to add the box or not, given the user's role and capabilities.